解决post报错的问题
