--- admin/src/main/java/com/admin/Token/tation/UserLoginToken.java
@@ -1,17 +1,0 @@
-package com.admin.Token.tation;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-/**
- * 用于登录后才能操作
- * @author MRC
- * @date 2019年4月4日 下午7:02:00
- */
-@Target({ElementType.METHOD, ElementType.TYPE})
-@Retention(RetentionPolicy.RUNTIME)
-public @interface UserLoginToken {
-    boolean required() default true;
-}
