--- admin/src/main/java/com/admin/admin/Dao/UserRoleDao.java
@@ -1,28 +1,0 @@
-package com.admin.admin.Dao;
-
-import com.admin.admin.Entity.Menu;
-import com.admin.admin.Entity.UserRole;
-import com.admin.model.ParamterModel;
-import com.admin.model.UserRoleModel;
-import org.apache.ibatis.annotations.Mapper;
-import org.apache.ibatis.annotations.Param;
-
-import java.util.List;
-
-@Mapper
-public interface UserRoleDao {
-
-    List<Menu> GetList(@Param("UserId")  int UserId);
-
-    //新增
-    int AddUserRole(  UserRole userRole);
-
-    //修改
-    int UpdateUserRole( UserRole userRole);
-    //删除
-    int DelUserRole(ParamterModel Paramter);
-    //获取当前用户的权限
-    List<UserRoleModel> GetRoleList(@Param("id") int id);
-
-
-}
