+++ admin/src/main/java/com/admin/admin/dao/UserPermissionGroupDao.java
@@ -1,0 +1,25 @@
+package com.admin.admin.dao;
+
+import com.admin.admin.entity.UserPermissionGroup;
+import com.admin.model.ParamterModel;
+import org.apache.ibatis.annotations.Mapper;
+import org.apache.ibatis.annotations.Param;
+
+import java.util.List;
+
+@Mapper
+public interface UserPermissionGroupDao {
+
+    //新增
+    int SaveUserGroup( UserPermissionGroup userGroup);
+    //修改
+    int UpdateGroup(UserPermissionGroup  userGroup);
+    //删除
+    int DeleteGroup(ParamterModel Paramter);
+
+    //获取
+    UserPermissionGroup GetGroup(@Param("id") int id);
+
+    //权限组列表
+    List<UserPermissionGroup> ListGroup(@Param("flag") boolean flag);
+}
