+++ admin/src/main/java/com/admin/admin/entity/UserRole.java
@@ -1,0 +1,64 @@
+package com.admin.admin.entity;
+
+import java.util.Date;
+
+public class UserRole {
+
+    private int role_id;
+    private String role_name;
+    private int menu_id;
+    private Date createtime;
+    private boolean status;
+    private int permission_id;
+
+    public int getRole_id() {
+        return role_id;
+    }
+
+    public void setRole_id(int role_id) {
+        this.role_id = role_id;
+    }
+
+    public String getRole_name() {
+        return role_name;
+    }
+
+    public void setRole_name(String role_name) {
+        this.role_name = role_name;
+    }
+
+    public int getMenu_id() {
+        return menu_id;
+    }
+
+    public UserRole setMenu_id(int menu_id) {
+        this.menu_id = menu_id;
+        return this;
+    }
+
+    public Date getCreatetime() {
+        return createtime;
+    }
+
+    public void setCreatetime(Date createtime) {
+        this.createtime = createtime;
+    }
+
+    public boolean isStatus() {
+        return status;
+    }
+
+    public void setStatus(boolean status) {
+        this.status = status;
+    }
+
+    public int getPermission_id() {
+        return permission_id;
+    }
+
+    public void setPermission_id(int permission_id) {
+        this.permission_id = permission_id;
+    }
+
+
+}
