+++ admin/src/main/java/com/admin/model/MenuData.java
@@ -1,0 +1,25 @@
+package com.admin.model;
+
+public class MenuData {
+    private String title;
+
+    public String getTitle() {
+        return title;
+    }
+
+    public void setTitle(String title) {
+        this.title = title;
+    }
+
+    public String getIcon() {
+        return icon;
+    }
+
+    public void setIcon(String icon) {
+        this.icon = icon;
+    }
+
+    private String icon;
+
+
+}
