--- admin/src/main/java/com/admin/model/Menudata.java
@@ -1,25 +1,0 @@
-package com.admin.model;
-
-public class Menudata {
-    private String title;
-
-    public String getTitle() {
-        return title;
-    }
-
-    public void setTitle(String title) {
-        this.title = title;
-    }
-
-    public String getIcon() {
-        return icon;
-    }
-
-    public void setIcon(String icon) {
-        this.icon = icon;
-    }
-
-    private String icon;
-
-
-}
