+++ admin/src/main/java/com/admin/token/tation/PassToken.java
@@ -1,0 +1,17 @@
+package com.admin.token.tation;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+/***
+ * 用来跳过验证的 PassToken
+ * @author MRC
+ * @date 2019年4月4日 下午7:01:25
+ */
+@Target({ElementType.METHOD, ElementType.TYPE})
+@Retention(RetentionPolicy.RUNTIME)
+public @interface PassToken {
+    boolean required() default true;
+}
