+++ admin/src/main/java/com/admin/token/tation/UserLoginToken.java
@@ -1,0 +1,17 @@
+package com.admin.token.tation;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+/**
+ * 用于登录后才能操作
+ * @author MRC
+ * @date 这个注解表示需要验证token
+ */
+@Target({ElementType.METHOD, ElementType.TYPE})
+@Retention(RetentionPolicy.RUNTIME)
+public @interface UserLoginToken {
+    boolean required() default true;
+}
