+++ person/src/main/java/com/person/person/PersonApplication.java
@@ -3,9 +3,11 @@ package com.person.person;
 import org.mybatis.spring.annotation.MapperScan;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.web.servlet.ServletComponentScan;
 
 @MapperScan("com.person.person.Personnel.Dao")
 @SpringBootApplication
+@ServletComponentScan //还有什么配置吗？ 我又来了
 public class PersonApplication {
 
     public static void main(String[] args) {
