+++ person/src/main/resources/application-dev.yml
@@ -1,5 +1,5 @@
 server:
-  port: 8088
+  port: 10086
 
 spring:
   datasource:
